scale_type_resolver::portable_registry

Function bits_from_metadata

pub fn bits_from_metadata(
    ty: &TypeDefBitSequence<PortableForm>,
    types: &PortableRegistry,
) -> Result<(BitsOrderFormat, BitsStoreFormat), Error>

Given some type information in the form of a scale_info::PortableRegistry, and a scale_info::TypeDefBitSequence, this returns information about the order and store format of the bit sequence.