scale_type_resolver

Module visitor

A concrete ResolvedTypeVisitor implementation that allows you to provide closures to configure it. Using this is often a lot easier than implementing ResolvedTypeVisitor yourself, but does require an additional dependency and may be a touch less performant.

Structs

ConcreteFieldIter

A concrete iterator over struct or variant fields information.

ConcreteResolvedTypeVisitor

This is a concrete visitor which implements ResolvedTypeVisitor. It’s instantiated by calling the standalone new function; see the docs there for more information.

Functions

new

Instantiate a new ConcreteResolvedTypeVisitor by providing a closure which is called by default if we don’t add any explicit handler for a given SCALE type.