scale_type_resolver::visitor

Function new

pub fn new<'resolver, Context, TypeId, Output, NewUnhandledFn>(
    context: Context,
    unhandled_fn: NewUnhandledFn,
) -> ConcreteResolvedTypeVisitor<'resolver, Context, TypeId, Output, NewUnhandledFn, impl FnOnce(Context) -> Output, impl FnOnce(Context, &mut dyn PathIter<'resolver>, &mut dyn FieldIter<'resolver, TypeId>) -> Output, impl FnOnce(Context, &mut dyn PathIter<'resolver>, &mut dyn VariantIter<'resolver, ConcreteFieldIter<'resolver, TypeId>>) -> Output, impl FnOnce(Context, &mut dyn PathIter<'resolver>, TypeId) -> Output, impl FnOnce(Context, TypeId, usize) -> Output, impl FnOnce(Context, &mut dyn ExactSizeIterator<Item = TypeId>) -> Output, impl FnOnce(Context, Primitive) -> Output, impl FnOnce(Context, TypeId) -> Output, impl FnOnce(Context, BitsStoreFormat, BitsOrderFormat) -> Output>
where
    NewUnhandledFn: FnOnce(Context, UnhandledKind) -> Output + Clone,

Instantiate a new ConcreteResolvedTypeVisitor by providing a closure which is called by default if we don’t add any explicit handler for a given SCALE type.

It’s expected that you’ll attach handlers for any types you’re interested in getting more details about like so:

use scale_type_resolver::{ TypeResolver, ResolvedTypeVisitor };

// Some dummy type that we're saying can resolve types:
struct MyTypeResolver;
impl TypeResolver for MyTypeResolver {
    type TypeId = u32;
    type Error = u8;
    fn resolve_type<'this, V: ResolvedTypeVisitor<'this, TypeId = Self::TypeId>>(
        &'this self,
        type_id: Self::TypeId,
        visitor: V,
    ) -> Result<V::Value, Self::Error> {
        Ok(visitor.visit_not_found())
    }
}

// Now, we can create a visitor using this `visitor::new` function.
// This has specific handling for composite and variant types, falling
// back to returning `1u64` if some other type was found.
let context = ();
let visitor = scale_type_resolver::visitor::new(context, |_context, _unhandled_kind| 1u64)
    .visit_composite(|_context, _composite_path, _composite_fields| 2)
    .visit_primitive(|_context, _primitive_type| 3);

// Now, when we provide the visitor to a type resolver, we'll get back a result
// containing either the value returned from the matched `visit_*` call above, or
// an error if the type resolver itself had an issue.
MyTypeResolver.resolve_type(123, visitor);

The visit_* methods provided each take closures which have a similar type signature to the underlying trait methods on ResolvedTypeVisitor, with small differences where necessary to avoid type and ownership issues. The first argument to every function is some arbitrary context which is provided as the first argument to crate::visitor::new().

Using this concrete visitor is expected to be almost as optimal as implementing the ResolvedTypeVisitor trait manually. One area where it makes a small sacrifice to this is in the ConcreteResolvedTypeVisitor::visit_variant() method, which must collect the variant fields into a smallvec::SmallVec to avoid ownership issues. This will only allocate if more than 16 variant fields exist. The default “unhandled” function here must also implement Clone, which isn’t necessary in a raw trait implementation, since it will, by default, be called in all of the other impls if no methods are provided.